Code to combine scores across regions in the same state

+++ motifs/scripts/qsub/scanMotifsOnClusters.sh
@@ -18,17 +18,27 @@ M using the -m option.
 OPTIONS:
    -h     Show this message and exit
    -m NUM Number of lines in split files.    
+   -r STR Path to a region file. Can be used to combine scores for 
+          regions of the input bed files.
+   -d STR Path where merged results will be written.
+          Default is OUTDIR/merged. 
 EOF
 }
 
 NLINES=1000
-while getopts "hn:" opt
+REGIONFILE=
+MERGEDIR=
+while getopts "hn:r:d:" opt
 do
     case $opt in
 	h)
 	    usage; exit;;
 	n)
 	    NLINES=$OPTARG;;
+	r) 
+	    REGIONFILE=$OPTARG;;
+	d) 
+	    MERGEDIR=$OPTARG;;
 	?)
 	    usage
             exit 1;;
@@ -45,6 +55,15 @@ OUTDIR=$2
 
 if [ ! -d ${OUTDIR}/tmp ]; then
     mkdir -p ${OUTDIR}/tmp
+fi
+
+if [ ! -z $REGIONFILE ]; then
+    if [ -z $MERGEDIR ]; then
+	MERGEDIR=${OUTDIR}/merged
+    fi
+    if [ ! -d ${MERGEDIR} ]; then
+	mkdir ${MERGEDIR}
+    fi
 fi
 
 HOMERSRC="${LABHOME}/software/homer/bin/"
@@ -64,9 +83,14 @@ while read -r bedfile; do
     countfile=${OUTDIR}/${pref}_vs_${suf}_counts.txt
 
     scorenpz=${OUTDIR}/${pref}_vs_${suf}_scores.npz
+    mergednpz=${MERGEDIR}/${pref}_vs_${suf}_merged_scores.npz
 
-    if [[ -f $scorenpz ]]; then
+    if [[ -f $scorenpz ]] && [[ -z $REGIONFILE ]] ; then
 	echo "Output file $scorenpz exists. Skipping" 1>&2
+	continue
+    fi
+    if [[ -f $mergednpz ]] && [[ ! -z $REGIONFILE ]] ; then
+	echo "Output file $mergednpz exists. Skipping" 1>&2
 	continue
     fi
 
@@ -77,23 +101,33 @@ while read -r bedfile; do
 	rm $countfile
     fi
 
-    touch $scorefile
-    touch $countfile
-
     echo "#!/bin/bash" > $script
     echo "module add python/2.7" >> $script
     echo "module add perl-scg" >> $script
     echo "PATH=${HOMERSRC}:${PATH}" >> $script
-    echo "zcat $bedfile | awk 'BEGIN{OFS=\"\t\"}{print NR,\$1,\$2,\$3,\"+\"}' | split -d -l $NLINES - $regionPref" >> $script
-    
-    echo "for regionFile in \`ls ${regionPref}*\`; do" >> $script
-    echo "    perl ${HOMERSRC}/annotatePeaks.pl \$regionFile hg19 -size given -noann -nogene -m $MOTFILE -mscore >> $scorefile" >> $script
-    echo "    perl ${HOMERSRC}/annotatePeaks.pl \$regionFile hg19 -size given -noann -nogene -m $MOTFILE -nmotifs >> $countfile" >> $script
-    echo "    rm \$regionFile" >> $script
-    echo "done" >> $script
 
-    echo "python ${SRCDIR}/python/mergeHomerAnnotate.py $OUTDIR ${pref}_vs_${suf} $scorenpz --insufs _scores.txt,_counts.txt" >> $script
-    echo "rm $scorefile $countfile" >> $script
+    if [ ! -f $scorenpz ]; then
+	touch $scorefile
+	touch $countfile
+	
+	echo "zcat $bedfile | awk 'BEGIN{OFS=\"\t\"}{print NR,\$1,\$2,\$3,\"+\"}' | split -d -l $NLINES - $regionPref" >> $script
+	echo "for regionFile in \`ls ${regionPref}*\`; do" >> $script
+	echo "    perl ${HOMERSRC}/annotatePeaks.pl \$regionFile hg19 -size given -noann -nogene -m $MOTFILE -mscore >> $scorefile" >> $script
+	echo "    perl ${HOMERSRC}/annotatePeaks.pl \$regionFile hg19 -size given -noann -nogene -m $MOTFILE -nmotifs >> $countfile" >> $script
+	echo "    rm \$regionFile" >> $script
+	echo "done" >> $script
+
+	echo "python ${SRCDIR}/python/mergeHomerAnnotate.py $OUTDIR ${pref}_vs_${suf} $scorenpz --insufs _scores.txt,_counts.txt" >> $script
+	echo "rm $scorefile $countfile" >> $script
+    fi
+    if [ ! -z $REGIONFILE ]; then
+	idfile=${MERGEDIR}/${pref}_vs_${suf}_region_ids.txt
+	echo "module add bedtools/2.18.0" >> $script
+	# Get the closest region for each element of the bed file. If there are multiple overlapping
+	# keep only one. If d > 0 (non-overlapping region), then output ".".
+	echo "closestBed -a $bedfile -b $REGIONFILE -d -t first | awk '{if(\$11==0){print \$10} else{print \".\"}}' > $idfile" >> $script
+	echo "python ${SRCDIR}/python/combineRegionScores.py $scorenpz $idfile $mergednpz" >> $script
+    fi
 
     qsub -q standard -N $pref -l h_vmem=4G -l h_rt=6:00:00 -e $errfile -o /dev/null $script
 done
